tuning bookmarks

+++ resources/functions/bookmarks.php
@@ -1,17 +1,42 @@
 <?php
 /**
+ * Get a list of the bookmarks of the currently logged in user.
+ * @return array
+ */
+function bibliographie_bookmarks_get_bookmarks () {
+	if(BIBLIOGRAPHIE_CACHING and file_exists(BIBLIOGRAPHIE_ROOT_PATH.'/cache/bookmarks_'.((int) bibliographie_user_get_id()).'.json'))
+		return json_decode(file_get_contents(BIBLIOGRAPHIE_ROOT_PATH.'/cache/bookmarks_'.((int) bibliographie_user_get_id()).'.json'));
+
+	$bookmarks = array();
+	$bookmarksResult = mysql_query("SELECT publications.`pub_id` FROM
+		`a2userbookmarklists` bookmarks,
+		`a2publication` publications
+	WHERE
+		publications.`pub_id` = bookmarks.`pub_id` AND
+		bookmarks.`user_id` = ".((int) bibliographie_user_get_id()."
+	ORDER BY
+		publications.`year` DESC"));
+	
+	if(mysql_num_rows($bookmarksResult) > 0)
+		while($bookmark = mysql_fetch_object($bookmarksResult))
+			$bookmarks[] = $bookmark->pub_id;
+
+	if(BIBLIOGRAPHIE_CACHING){
+		$cacheFile = fopen(BIBLIOGRAPHIE_ROOT_PATH.'/cache/bookmarks_'.((int) bibliographie_user_get_id()).'.json', 'w+');
+		fwrite($cacheFile, json_encode($bookmarks));
+		fclose($cacheFile);
+	}
+
+	return $bookmarks;
+}
+
+/**
  * Set a bookmark for a publication for the logged in user.
  * @param int $pub_id
  * @return bool
  */
 function bibliographie_bookmarks_set_bookmark ($pub_id) {
-	return mysql_query("INSERT INTO `a2userbookmarklists` (
-	`user_id`,
-	`pub_id`
-) VALUES (
-	".((int) bibliographie_user_get_id()).",
-	".((int) $pub_id)."
-)");
+	return (bool) bibliographie_bookmarks_set_bookmarks_for_list(array($pub_id));
 }
 
 /**
@@ -20,8 +45,7 @@ function bibliographie_bookmarks_set_bookmark ($pub_id) {
  * @return bool
  */
 function bibliographie_bookmarks_unset_bookmark ($pub_id) {
-	mysql_query("DELETE FROM `a2userbookmarklists` WHERE `user_id` = ".((int) bibliographie_user_get_id())." AND `pub_id` = ".((int) $pub_id));
-	return (bool) mysql_affected_rows();
+	return (bool) bibliographie_bookmarks_unset_bookmarks_for_list(array($pub_id));
 }
 
 /**
@@ -29,8 +53,20 @@ function bibliographie_bookmarks_unset_bookmark ($pub_id) {
  * @return bool
  */
 function bibliographie_bookmarks_clear_bookmarks () {
-	mysql_query("DELETE FROM `a2userbookmarklists` WHERE `user_id` = ".((int) bibliographie_user_get_id()));
-	return (bool) mysql_affected_rows();
+	if(count(bibliographie_bookmarks_get_bookmarks()) > 0){
+		mysql_query("DELETE FROM `a2userbookmarklists` WHERE `user_id` = ".((int) bibliographie_user_get_id()));
+		$return = mysql_affected_rows();
+
+		if($return){
+			$cacheFile = fopen(BIBLIOGRAPHIE_ROOT_PATH.'/cache/bookmarks_'.((int) bibliographie_user_get_id()).'.json', 'w+');
+			fwrite($cacheFile, '[]');
+			fclose($cacheFile);
+		}
+
+		return $return;
+	}
+
+	return false;
 }
 
 /**
@@ -38,7 +74,7 @@ function bibliographie_bookmarks_clear_bookmarks () {
  * @param int $pub_id
  */
 function bibliographie_bookmarks_check_publication ($pub_id) {
-	return (bool) mysql_num_rows(mysql_query("SELECT * FROM `a2userbookmarklists` WHERE `user_id` = ".((int) bibliographie_user_get_id())." AND `pub_id` = ".((int) $pub_id)));
+	return in_array($pub_id, bibliographie_bookmarks_get_bookmarks());
 }
 
 /**
@@ -97,4 +133,68 @@ function bibliographie_bookmarks_unset_bookmark (pub_id) {
 	/* ]]> */
 </script>
 <?php
+}
+
+/**
+ * Set bookmarks for a given list of publications.
+ * @param array $list
+ * @return mixed Amount of set bookmarks or false on error.
+ */
+function bibliographie_bookmarks_set_bookmarks_for_list (array $list) {
+	if(count($list) > 0){
+		$mysqlString = "";
+
+		$list = array_diff($list, bibliographie_bookmarks_get_bookmarks());
+		if(count($list) > 0)
+			foreach($list as $pub_id){
+				if(!empty($mysqlString))
+					$mysqlString .= ",";
+				$mysqlString .= "(".((int) bibliographie_user_get_id()).",".((int) $pub_id).")";
+			}
+
+		$return = 0;
+		if(!empty($mysqlString)){
+			mysql_query("INSERT INTO `a2userbookmarklists` (`user_id`,`pub_id`) VALUES ".$mysqlString.";");
+			$return = mysql_affected_rows();
+		}
+
+		if($return > 0)
+			bibliographie_purge_cache('bookmarks_'.((int) bibliographie_user_get_id()));
+
+		return $return;
+	}
+
+	return false;
+}
+
+/**
+ * Unset bookmarks for a given list of publications
+ * @param array $list
+ * @return mixed Amount of unset bookmarks or false on error.
+ */
+function bibliographie_bookmarks_unset_bookmarks_for_list (array $list) {
+	if(count($list)){
+		$mysqlString = "";
+
+		$list = array_intersect($list, bibliographie_bookmarks_get_bookmarks());
+		if(count($list) > 0)
+			foreach($list as $pub_id){
+				if(!empty($mysqlString))
+					$mysqlString .= " OR ";
+				$mysqlString .= "`pub_id` = ".((int) $pub_id);
+			}
+
+		$return = 0;
+		if(!empty($mysqlString)){
+			mysql_query("DELETE FROM `a2userbookmarklists` WHERE `user_id` = ".((int) bibliographie_user_get_id())." AND (".$mysqlString.")");
+			$return = mysql_affected_rows();
+		}
+
+		if($return > 0)
+			bibliographie_purge_cache('bookmarks_'.((int) bibliographie_user_get_id()));
+
+		return $return;
+	}
+
+	return false;
 }
